support zfs zvols as volumes
(cherry picked from commit c3173c347753b52e32b6f77b2952086dfb7a127e)

+++ pkg/pillar/cmd/volumemgr/create.go
@@ -12,6 +12,7 @@ import (
 	"github.com/lf-edge/eve/pkg/pillar/cas"
 	"github.com/lf-edge/eve/pkg/pillar/diskmetrics"
 	"github.com/lf-edge/eve/pkg/pillar/types"
+	"github.com/lf-edge/eve/pkg/pillar/zfs"
 )
 
 // createVolume does not update status but returns
@@ -33,55 +34,82 @@ func createVdiskVolume(ctx *volumemgrContext, status types.VolumeStatus,
 
 	created := false
 
+	persistFsType := ctx.persistType
+
 	// this is the target location, where we expect the volume to be
 	filelocation := status.PathName()
-	if _, err := os.Stat(filelocation); err == nil {
-		errStr := fmt.Sprintf("Can not create %s for %s: exists",
-			filelocation, status.Key())
-		log.Error(errStr)
-		return created, "", errors.New(errStr)
+
+	if persistFsType != types.PersistZFS {
+		if _, err := os.Stat(filelocation); err == nil {
+			errStr := fmt.Sprintf("Can not create %s for %s: exists",
+				filelocation, status.Key())
+			log.Error(errStr)
+			return created, "", errors.New(errStr)
+		}
 	}
 
-	// use the edge-containers library to extract the data we need
-	puller := registry.Puller{
-		Image: ref,
-	}
+	switch persistFsType {
+	case types.PersistZFS:
+		pathToFile, err := getVolumeFilePath(ctx, status)
+		if err != nil {
+			errStr := fmt.Sprintf("Error obtaining file for zvol at volume %s, error=%v",
+				status.Key(), err)
+			log.Error(errStr)
+			return created, "", errors.New(errStr)
+		}
+		zVolName := status.ZVolName(types.VolumeZFSPool)
+		zVolDevice := zfs.GetZVolDeviceByDataset(zVolName)
+		if zVolDevice == "" {
+			errStr := fmt.Sprintf("Error finding zfs zvol %s", zVolName)
+			log.Error(errStr)
+			return created, "", errors.New(errStr)
+		}
+		if err := diskmetrics.ConvertImg(log, pathToFile, zVolDevice, "raw"); err != nil {
+			errStr := fmt.Sprintf("Error converting %s to zfs zvol %s: %v",
+				pathToFile, zVolDevice, err)
+			log.Error(errStr)
+			return created, "", errors.New(errStr)
+		}
+		filelocation = zVolDevice
+	default:
+		// use the edge-containers library to extract the data we need
+		puller := registry.Puller{
+			Image: ref,
+		}
 
-	casClient, err := cas.NewCAS(casClientType)
-	if err != nil {
-		err = fmt.Errorf("Run: exception while initializing CAS client: %s", err.Error())
-		return created, "", err
-	}
-	defer casClient.CloseClient()
-	ctrdCtx, done := casClient.CtrNewUserServicesCtx()
-	defer done()
+		casClient, err := cas.NewCAS(casClientType)
+		if err != nil {
+			err = fmt.Errorf("Run: exception while initializing CAS client: %s", err.Error())
+			return created, "", err
+		}
+		defer casClient.CloseClient()
+		ctrdCtx, done := casClient.CtrNewUserServicesCtx()
+		defer done()
 
-	resolver, err := casClient.Resolver(ctrdCtx)
-	if err != nil {
-		errStr := fmt.Sprintf("error getting CAS resolver: %v", err)
-		log.Error(errStr)
-		return created, "", errors.New(errStr)
-	}
-
-	// create a writer for the file where we want
-	f, err := os.Create(filelocation)
-	if err != nil {
-		errStr := fmt.Sprintf("error creating target file at %s: %v", filelocation, err)
-		log.Error(errStr)
-		return created, "", errors.New(errStr)
-	}
-	defer f.Close()
-
-	if _, _, err := puller.Pull(&registry.FilesTarget{Root: f, AcceptHash: true}, 0, false, os.Stderr, resolver); err != nil {
-		errStr := fmt.Sprintf("error pulling %s from containerd: %v", ref, err)
-		log.Error(errStr)
-		return created, "", errors.New(errStr)
-	}
-
-	// Do we need to expand disk?
-	if err := maybeResizeDisk(filelocation, status.MaxVolSize); err != nil {
-		log.Error(err)
-		return created, "", err
+		resolver, err := casClient.Resolver(ctrdCtx)
+		if err != nil {
+			errStr := fmt.Sprintf("error getting CAS resolver: %v", err)
+			log.Error(errStr)
+			return created, "", errors.New(errStr)
+		}
+		// create a writer for the file where we want
+		f, err := os.Create(filelocation)
+		if err != nil {
+			errStr := fmt.Sprintf("error creating target file at %s: %v", filelocation, err)
+			log.Error(errStr)
+			return created, "", errors.New(errStr)
+		}
+		defer f.Close()
+		if _, _, err := puller.Pull(&registry.FilesTarget{Root: f, AcceptHash: true}, 0, false, os.Stderr, resolver); err != nil {
+			errStr := fmt.Sprintf("error pulling %s from containerd: %v", ref, err)
+			log.Error(errStr)
+			return created, "", errors.New(errStr)
+		}
+		// Do we need to expand disk?
+		if err := maybeResizeDisk(filelocation, status.MaxVolSize); err != nil {
+			log.Error(err)
+			return created, "", err
+		}
 	}
 
 	log.Functionf("Extract DONE from %s to %s", ref, filelocation)
@@ -126,8 +154,9 @@ func createContainerVolume(ctx *volumemgrContext, status types.VolumeStatus,
 func destroyVolume(ctx *volumemgrContext, status types.VolumeStatus) (bool, string, error) {
 
 	log.Functionf("destroyVolume(%s)", status.Key())
-	if !status.VolumeCreated {
-		log.Functionf("destroyVolume(%s) nothing was created", status.Key())
+	// we have no explicit un-prepare action for now, so it works with prepared or created volumes
+	if status.SubState != types.VolumeSubStateCreated && status.SubState != types.VolumeSubStatePrepareDone {
+		log.Functionf("destroyVolume(%s) nothing was created/prepared", status.Key())
 		return false, status.FileLocation, nil
 	}
 
@@ -152,13 +181,30 @@ func destroyVolume(ctx *volumemgrContext, status types.VolumeStatus) (bool, stri
 // new values for VolumeCreated, FileLocation, and error
 func destroyVdiskVolume(ctx *volumemgrContext, status types.VolumeStatus) (bool, string, error) {
 
-	created := status.VolumeCreated
+	created := status.SubState == types.VolumeSubStateCreated
 	filelocation := status.FileLocation
 	log.Functionf("Delete copy at %s", filelocation)
-	if err := os.RemoveAll(filelocation); err != nil {
-		log.Error(err)
-		filelocation = ""
-		return created, filelocation, err
+
+	info, err := os.Stat(filelocation)
+	if err != nil {
+		errStr := fmt.Sprintf("Error get stat of file %s: %v",
+			filelocation, err)
+		return created, "", errors.New(errStr)
+	}
+	if info.Mode()&os.ModeDevice != 0 {
+		//Assume this is zfs device
+		zVolName := status.ZVolName(types.VolumeZFSPool)
+		if stdoutStderr, err := zfs.DestroyDataset(log, zVolName); err != nil {
+			errStr := fmt.Sprintf("Error destroying zfs zvol at %s, error=%v, output=%s",
+				zVolName, err, stdoutStderr)
+			log.Error(errStr)
+			return created, "", errors.New(errStr)
+		}
+	} else {
+		if err := os.RemoveAll(filelocation); err != nil {
+			log.Error(err)
+			return created, "", err
+		}
 	}
 	filelocation = ""
 	created = false
@@ -170,7 +216,7 @@ func destroyVdiskVolume(ctx *volumemgrContext, status types.VolumeStatus) (bool,
 // new values for VolumeCreated, FileLocation, and error
 func destroyContainerVolume(ctx *volumemgrContext, status types.VolumeStatus) (bool, string, error) {
 
-	created := status.VolumeCreated
+	created := status.SubState == types.VolumeSubStateCreated
 	filelocation := status.FileLocation
 	log.Functionf("Removing container volume %s", filelocation)
 	if err := ctx.casClient.RemoveContainerRootDir(filelocation); err != nil {
@@ -182,22 +228,35 @@ func destroyContainerVolume(ctx *volumemgrContext, status types.VolumeStatus) (b
 	return created, filelocation, nil
 }
 
+// returns size and indicates do we need to resize disk to be at least maxsizebytes
+func checkResizeDisk(diskfile string, maxsizebytes uint64) (uint64, bool, error) {
+	vSize, err := diskmetrics.GetDiskVirtualSize(log, diskfile)
+	if err != nil {
+		return 0, false, err
+	}
+	if vSize > maxsizebytes {
+		log.Warnf("Virtual size (%d) of provided volume(%s) is larger than provided MaxVolSize (%d). "+
+			"Will use virtual size.", vSize, diskfile, maxsizebytes)
+		return vSize, false, nil
+	}
+	return maxsizebytes, vSize != maxsizebytes, nil
+}
+
 // Make sure the (virtual) size of the disk is at least maxsizebytes
 func maybeResizeDisk(diskfile string, maxsizebytes uint64) error {
 	if maxsizebytes == 0 {
 		return nil
 	}
-	currentSize, err := diskmetrics.GetDiskVirtualSize(log, diskfile)
+	log.Functionf("maybeResizeDisk(%s) current to %d",
+		diskfile, maxsizebytes)
+	size, resize, err := checkResizeDisk(diskfile, maxsizebytes)
 	if err != nil {
-		return err
+		return fmt.Errorf("maybeResizeDisk checkResizeDisk error: %s", err)
 	}
-	log.Functionf("maybeResizeDisk(%s) current %d to %d",
-		diskfile, currentSize, maxsizebytes)
-	if maxsizebytes < currentSize {
-		log.Warnf("maybeResizeDisk(%s) already above maxsize  %d vs. %d",
-			diskfile, maxsizebytes, currentSize)
-		return nil
+	if resize {
+		log.Functionf("maybeResizeDisk(%s) resize to %d",
+			diskfile, size)
+		return diskmetrics.ResizeImg(log, diskfile, size)
 	}
-	err = diskmetrics.ResizeImg(log, diskfile, maxsizebytes)
-	return err
+	return nil
 }
